legal: rewrite ToS + Privacy (Quebec Law 25), redesign contact page, ToS acceptance audit log, UI polish
- Terms of Service: full rewrite (23 sections), Quebec governing law, age 14 min, CAD  liability cap, one-time purchase model, third-party disclosures (Clerk, OpenAI, Stripe, Render, S3)
- Privacy Policy: full rewrite (20 sections), Quebec Law 25 compliance, Amazon S3 added, Sentry removed, generalized data retention language, Privacy Officer contact
- Contact page: redesigned from dual-card to single email CTA with topic tiles (billing, privacy, technical, general), FAQ aligned with updated ToS/Privacy
- Footer: removed firm name mention, simplified to 'All rights reserved'
- ToS acceptance gate: bumped version to 2026-05-14, updated summary bullets (age 14, Law 25, voice recording, one-time purchases), re-acceptance required for all existing users
- ToS audit log: new tos_acceptances table (append-only), captures IP + User-Agent per acceptance event, migration 0019, backend config + model + endpoint updated
- sitemap.ts: updated lastModified for /terms and /privacy to 2026-05-14
- TaskUpsellCard: removed 'One-time purchase' subtitle, enlarged title to match price, amber icon
- RecentAttemptsFeed: aligned style with dashboard table (plain score badge, gap-12 column alignment, matching header)
- RecentAttemptsFeed: renamed 'Report' to 'View', added ?from=/progress param
- ReportPage: dynamic back link reads ?from= param (progress/history/dashboard), Suspense wrapper added
- CountdownOverlay: removed 'Begin speaking now' subtitle, GO! styled amber

+++ apps/api/app/schemas/history.py
@@ -21,7 +21,7 @@ class HistoryItem(BaseModel):
     task_title: str               # human-readable prompt title
     is_mock_test: bool
     status: str                   # "pending" | "processing" | "complete" | "failed"
-    estimated_band: float | None  # None if not yet scored
+    estimated_band: int | None       # None if not yet scored; whole number 1–12
     created_at: datetime
 
 
@@ -40,7 +40,7 @@ class MockExamTaskResult(BaseModel):
     """Per-task band score within one mock exam session."""
     task_number:    int
     status:         str            # pending | processing | complete | failed
-    estimated_band: float | None
+    estimated_band: int | None      # whole number 1–12, or None if not scored
 
 
 class MockExamSession(BaseModel):
@@ -75,7 +75,7 @@ class DimensionScorePoint(BaseModel):
 class TaskScorePoint(BaseModel):
     """One historical band score for a specific skill + task_number."""
     attempt_id:     UUID
-    estimated_band: float
+    estimated_band: int             # whole number 1–12
     completed_at:   datetime
     dimensions:     list[DimensionScorePoint] = []   # empty for legacy rows
